s22

+++ Assets/Scripts/Runtime/Systems/Entities Persistence/Editor/SaveManagerEditor.cs
@@ -13,14 +13,10 @@ namespace Universal.Runtime.Systems.EntitiesPersistence
 
             DrawDefaultInspector();
 
-            if (GUILayout.Button("New Game"))
-                saveLoadSystem.NewGame();
-            if (GUILayout.Button("Save Game"))
-                saveLoadSystem.SaveGame();
-            if (GUILayout.Button("Load Game"))
-                saveLoadSystem.LoadGame(gameName);
-            if (GUILayout.Button("Delete Game"))
-                saveLoadSystem.DeleteGame(gameName);
+            if (GUILayout.Button("New Game")) saveLoadSystem.NewGame();
+            if (GUILayout.Button("Save Game")) saveLoadSystem.SaveGame();
+            if (GUILayout.Button("Load Game")) saveLoadSystem.LoadGame(gameName);
+            if (GUILayout.Button("Delete Game")) saveLoadSystem.DeleteGame(gameName);
         }
     }
 }
